+++ src/xyz/oihana/schema/constants/traits/AuthTrait.php
@@ -8,6 +8,7 @@ use xyz\oihana\schema\constants\traits\auth\KeyfileTrait;
 use xyz\oihana\schema\constants\traits\auth\PermissionTrait;
 use xyz\oihana\schema\constants\traits\auth\PolicyTrait;
 use xyz\oihana\schema\constants\traits\auth\RoleTrait;
+use xyz\oihana\schema\constants\traits\auth\ServiceTrait;
 use xyz\oihana\schema\constants\traits\auth\SessionTrait;
 use xyz\oihana\schema\constants\traits\auth\UserTrait;
 use xyz\oihana\schema\constants\traits\auth\WebApplicationTrait;
@@ -21,6 +22,7 @@ trait AuthTrait
         PermissionTrait ,
         PolicyTrait ,
         RoleTrait ,
+        ServiceTrait ,
         SessionTrait ,
         UserTrait ,
         WebApplicationTrait ,