+++ src/xyz/oihana/schema/constants/traits/auth/ServicesTrait.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\auth;
+
+/**
+ * The enumeration of all services properties.
+ *
+ * @package oihana\schema\constants\traits\auth
+ * @author  Marc Alcaraz
+ */
+trait ServicesTrait
+{
+    const string SERVICES       = 'services' ;
+    const string SERVICES_COUNT = 'servicesCount' ;
+}